pyproject.toml
[project]
name = "entity-relationship-extractor"
version = "0.1.0"
description = ""
authors = [
    {name = "Kyle Williams",email = "kyle.williams2k15@gmail.com"}
]
include = "app"
packages = [{ include = "app" }]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "spacy (>=3.8.5,<4.0.0)",
    "langchain (>=0.3.15,<0.4.0)",
    "langchain-anthropic (>=0.3.7,<0.4.0)",
    "markdown-it-py (>=3.0.0,<4.0.0)",
    "pydantic-ai (>=0.1.3,<0.2.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
